only create archive-schedule for production enviornments

+++ helpers/k8up-initiate-archive.sh
@@ -31,6 +31,7 @@ PROJECT=$(echo "$CONFIGMAP" | jq -r '.data.LAGOON_PROJECT')
 SAFE_PROJECT=$(echo "$CONFIGMAP" | jq -r '.data.LAGOON_SAFE_PROJECT')
 BRANCH=$(echo "$CONFIGMAP" | jq -r '.data.LAGOON_GIT_BRANCH')
 SAFE_BRANCH=$(echo "$CONFIGMAP" | jq -r '.data.LAGOON_GIT_SAFE_BRANCH')
+ENVIRONMENT_TYPE=$(echo "$CONFIGMAP" | jq -r '.data.LAGOON_ENVIRONMENT_TYPE')
 LAGOON_GIT_SHA="00000000000000000000000000000000000000000"
 OPENSHIFT_REGISTRY="docker-registry.default.svc:5000"
 ROUTER_URL=""
@@ -41,19 +42,25 @@ if oc get customresourcedefinition schedules.backup.appuio.ch > /dev/null; then
 
   # create archive only if there is a backup-schedule already existing for this project
   if oc -n ${OPENSHIFT_PROJECT} get schedule backup-schedule &> /dev/null; then
-    TEMPLATE_PARAMETERS=()
 
-    # Run Archive on Monday at 0300-0600
-    ARCHIVE_SCHEDULE=$( $(git rev-parse --show-toplevel)/images/oc-build-deploy-dind/scripts/convert-crontab.sh "${OPENSHIFT_PROJECT}" "M H(6-12) 2 * *")
-    TEMPLATE_PARAMETERS+=(-p ARCHIVE_SCHEDULE="${ARCHIVE_SCHEDULE}")
+    # create archive only if this is a production environment
+    if [[ "${ENVIRONMENT_TYPE}" == "production" ]]; then
+      TEMPLATE_PARAMETERS=()
 
-    TEMPLATE_PARAMETERS+=(-p ARCHIVE_BUCKET="${ARCHIVE_BUCKET}")
+      # Run Archive on Monday at 0300-0600
+      ARCHIVE_SCHEDULE=$( $(git rev-parse --show-toplevel)/images/oc-build-deploy-dind/scripts/convert-crontab.sh "${OPENSHIFT_PROJECT}" "M H(6-12) 2 * *")
+      TEMPLATE_PARAMETERS+=(-p ARCHIVE_SCHEDULE="${ARCHIVE_SCHEDULE}")
 
-    OPENSHIFT_TEMPLATE="$(git rev-parse --show-toplevel)/images/oc-build-deploy-dind/openshift-templates/backup-archive-schedule.yml"
-    .  $(git rev-parse --show-toplevel)/images/oc-build-deploy-dind/scripts/exec-openshift-resources.sh
+      TEMPLATE_PARAMETERS+=(-p ARCHIVE_BUCKET="${ARCHIVE_BUCKET}")
 
-    oc apply -n ${OPENSHIFT_PROJECT} -f /tmp/k8up-archive-initiate.yml
-    rm /tmp/k8up-archive-initiate.yml
+      OPENSHIFT_TEMPLATE="$(git rev-parse --show-toplevel)/images/oc-build-deploy-dind/openshift-templates/backup-archive-schedule.yml"
+      .  $(git rev-parse --show-toplevel)/images/oc-build-deploy-dind/scripts/exec-openshift-resources.sh
+
+      oc apply -n ${OPENSHIFT_PROJECT} -f /tmp/k8up-archive-initiate.yml
+      rm /tmp/k8up-archive-initiate.yml
+    else
+      echo "${OPENSHIFT_PROJECT}: Not production environment, not creating an archive-schedule"
+    fi
   else
     echo "${OPENSHIFT_PROJECT}: No backup-schedule found for project, not creating an archive-schedule"
   fi
